function breakouts, typing, tag parsing
Updated several items in this file. Added a generate_parsed_post_list
function to breakout that functionality. The reason for this was it
made it easier to test. I also rearranged several functions to match the
flow of the 'generate_posts' function. That way when you are reading the
code it is easier to logically follow.

Added some minor type hints. More to come.

Tag parsing is properly implemented. Tags are now a list of strings
rather than just a string.