Adds date sorting/formatting, New classes for Metadata/Post Collections
Added the functionality to sort by date with metadata. This is useful
for sorting at the server level and not having to deal with parsing
dates in javascript on the front end. Now if you create a table or list
with metadata it will by default be in descending order.

To implement I needed to break out the PostCollection class into two sub
classes; MetadataCollection and PostsCollection. The main reason for
implementing these two sub classes rather than implementing the logic in
the PostsCollection class is that I want PostsCollection to be the top
level class exposed to the user. It should be extremely clean and only
present the metadata and posts attributes to the end user.