black linting, updates tests
Black linter added.

Adds fixtures for a couple items that allow reuse. Not sure if fixtures
are the correct way to go about testing but they seem to be working okay
for now. Changes import scheme which allowed me to fix the issue I was
having with __all__ in posts.py.